JSON parsing compatible with python 2 and 3
In python2, which ships out of the box with macOS, `print` is not a function.  But in python3, which many developers use `print` is a function.

python2 example `print "Hello World"`
python3 example `print("Hellow World")

To fix so that the shell launcher works both in python2 and python3, we add the `print_function` class from the `__future__` library.

This has been tested on python2 and python3.