Support for either Py2 or Py3. Dropped “OSX” label
No longer using the term “OSX” as that is outdated — instead all instances changed to macOS.

The latest python installation uses `python3` as the default command in the terminal.  The launcher would fail when the `python` command was not present even with `python3` loaded.  The script now “sniffs” for the installed Python interpreter command.  Users should not have to do anything except to ensure Python is installed.